refactor: activity 관련 manifest 코드를 presentation 모듈의 manifest로 이동, 예제 데이터를 Response 객체에 담아서 받음, named argument 사용, imp -> impl로 클래스명 변경
