+++ app/src/main/java/com/shypolarbear/android/di/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package com.shypolarbear.android.di
 
-import com.shypolarbear.data.repositoryimp.ExampleRepoImp
+import com.shypolarbear.data.repositoryimpl.ExampleRepoImpl
 import com.shypolarbear.domain.repository.ExampleRepo
 import dagger.Binds
 import dagger.Module
@@ -12,5 +12,5 @@ import dagger.hilt.components.SingletonComponent
 abstract class RepositoryModule {
 
     @Binds
-    abstract fun bindExampleRepo(repoImp: ExampleRepoImp): ExampleRepo
+    abstract fun bindExampleRepo(repoImp: ExampleRepoImpl): ExampleRepo
 }
