+++ data/src/main/java/com/shypolarbear/data/api/ExampleApi.kt
@@ -1,9 +1,10 @@
 package com.shypolarbear.data.api
 
 import com.shypolarbear.domain.model.ExampleModel
+import retrofit2.Response
 import retrofit2.http.GET
 
 interface ExampleApi {
     @GET("jokes/categories")
-    suspend fun getExample(): ExampleModel
+    suspend fun getExample(): Response<ExampleModel>
 }
