+++ presentation/src/main/AndroidManifest.xml
@@ -1,4 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+    <uses-permission android:name="android.permission.INTERNET" />
 
+    <application
+        android:allowBackup="true"
+        android:dataExtractionRules="@xml/data_extraction_rules"
+        android:fullBackupContent="@xml/backup_rules"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:theme="@style/Theme.SplashScreen.ShyPolarBear"
+        tools:targetApi="31">
+
+        <!--        <activity-->
+        <!--            android:name="com.shypolarbear.presentation.MainActivity"-->
+        <!--            android:exported="true"-->
+        <!--            android:label="@string/app_name"-->
+        <!--            android:theme="@style/Theme.ShyPolarBear">-->
+        <!--            <intent-filter>-->
+        <!--                <action android:name="android.intent.action.MAIN"/>-->
+        <!--                <category android:name="android.intent.category.LAUNCHER"/>-->
+        <!--            </intent-filter>-->
+        <!--        </activity>-->
+        <activity
+            android:name="com.shypolarbear.presentation.sample.SampleActivity"
+            android:exported="true"
+            android:label="@string/app_name">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
+                <category android:name="android.intent.category.LAUNCHER"/>
+            </intent-filter>
+        </activity>
+
+    </application>
 </manifest>
+
